interactwel/requirements.txt
Django==2.2.1
django-webpack-loader==0.6.0
pytz==2018.7
djangorestframework==3.8.2
django-filter>=2.1.0, <2.2
markdown==2.6.11
django-mysql-3.1.0==3.1.0
jsonfield==2.0.2
